+++ main.go
@@ -67,6 +67,8 @@ func main() {
 			cmd.DoctorParse(os.Args[2:])
 		case "migrate":
 			cmd.MigrateParse(os.Args[2:], clogger)
+		case "admin":
+			cmd.AdminParse(os.Args[2:], clogger)
 		}
 	}